Move the HMAC auth policy to an internal package
No functional changes.
