+++ sdk/data/azappconfig/internal/auth/policy_hmac_auth_parse_test.go
@@ -4,7 +4,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package azappconfig
+package auth
 
 import (
 	"testing"
@@ -13,7 +13,7 @@ import (
 )
 
 func TestHmacAuthParseConnectionString(t *testing.T) {
-	ep, id, sc, err := parseConnectionString("Endpoint=xX;Id=yY;Secret=ZmZm")
+	ep, id, sc, err := ParseConnectionString("Endpoint=xX;Id=yY;Secret=ZmZm")
 	require.NoError(t, err)
 	require.Equal(t, "xX", ep)
 	require.Equal(t, "yY", id)
@@ -25,7 +25,7 @@ func TestHmacAuthParseConnectionString(t *testing.T) {
 }
 
 func TestHmacAuthParseConnectionStringMixedOrder(t *testing.T) {
-	ep, id, sc, err := parseConnectionString("Id=yY;Secret=ZmZm;Endpoint=xX")
+	ep, id, sc, err := ParseConnectionString("Id=yY;Secret=ZmZm;Endpoint=xX")
 	require.NoError(t, err)
 	require.Equal(t, "xX", ep)
 	require.Equal(t, "yY", id)
@@ -37,7 +37,7 @@ func TestHmacAuthParseConnectionStringMixedOrder(t *testing.T) {
 }
 
 func TestHmacAuthParseConnectionStringExtraProperties(t *testing.T) {
-	ep, id, sc, err := parseConnectionString("A=aA;Endpoint=xX;B=bB;Id=yY;C=cC;Secret=ZmZm;D=dD;")
+	ep, id, sc, err := ParseConnectionString("A=aA;Endpoint=xX;B=bB;Id=yY;C=cC;Secret=ZmZm;D=dD;")
 	require.NoError(t, err)
 	require.Equal(t, "xX", ep)
 	require.Equal(t, "yY", id)
@@ -49,31 +49,31 @@ func TestHmacAuthParseConnectionStringExtraProperties(t *testing.T) {
 }
 
 func TestHmacAuthParseConnectionStringMissingEndoint(t *testing.T) {
-	_, _, _, err := parseConnectionString("Id=yY;Secret=ZmZm")
+	_, _, _, err := ParseConnectionString("Id=yY;Secret=ZmZm")
 	require.Error(t, err)
 }
 
 func TestHmacAuthParseConnectionStringMissingId(t *testing.T) {
-	_, _, _, err := parseConnectionString("Endpoint=xX;Secret=ZmZm")
+	_, _, _, err := ParseConnectionString("Endpoint=xX;Secret=ZmZm")
 	require.Error(t, err)
 }
 
 func TestHmacAuthParseConnectionStringMissingSecret(t *testing.T) {
-	_, _, _, err := parseConnectionString("Endpoint=xX;Id=yY")
+	_, _, _, err := ParseConnectionString("Endpoint=xX;Id=yY")
 	require.Error(t, err)
 }
 
 func TestHmacAuthParseConnectionStringDuplicateEndoint(t *testing.T) {
-	_, _, _, err := parseConnectionString("Endpoint=xX;Endpoint=xX;Id=yY;Secret=ZmZm")
+	_, _, _, err := ParseConnectionString("Endpoint=xX;Endpoint=xX;Id=yY;Secret=ZmZm")
 	require.Error(t, err)
 }
 
 func TestHmacAuthParseConnectionStringDuplicateId(t *testing.T) {
-	_, _, _, err := parseConnectionString("Endpoint=xX;Id=yY;Id=yY;Secret=ZmZm")
+	_, _, _, err := ParseConnectionString("Endpoint=xX;Id=yY;Id=yY;Secret=ZmZm")
 	require.Error(t, err)
 }
 
 func TestHmacAuthParseConnectionStringDuplicateSecret(t *testing.T) {
-	_, _, _, err := parseConnectionString("Endpoint=xX;Id=yY;Secret=ZmZm;Secret=zZ")
+	_, _, _, err := ParseConnectionString("Endpoint=xX;Id=yY;Secret=ZmZm;Secret=zZ")
 	require.Error(t, err)
 }
